Chore: Organize Module routes for cleanliness
