+++ src/routes/module-routes/module.asset.routes.tsx
