+++ src/routes/module-routes/module.inventory.routes.tsx
