+++ src/routes/module-routes/module.pos.routes.tsx
